Add einsum, polynomial, advanced indexing, N-D convolutions, optimization, and expanded indexing operators to pkgs/ndarray
TAG=agy
CONV=dc2285a9-9869-4336-8a63-5711443f7182

+++ pkgs/ndarray/lib/src/operations/distance.dart
@@ -325,7 +325,7 @@ NDArray<Float64> _pdistCosine<T extends Object>(
     );
 
     final NDArray<Float64> div = divide(dot, denom);
-    final one = NDArray<Float64>.fromList([1.0], [1], DType.float64);
+    final one = NDArray<Float64>.fromList([Float64(1.0)], [1], DType.float64);
     final NDArray<Float64> cosDistMatrix = subtract(one, div);
 
     final flatData = cosDistMatrix.data as Float64List;
@@ -384,7 +384,7 @@ NDArray<Float64> _cdistCosine<Ta extends Object, Tb extends Object>(
     );
 
     final NDArray<Float64> div = divide(dot, denom);
-    final one = NDArray<Float64>.fromList([1.0], [1], DType.float64);
+    final one = NDArray<Float64>.fromList([Float64(1.0)], [1], DType.float64);
     subtract(one, div, out: result);
 
     if (out != null) {
